+++ src/scripts/three/lights.js
@@ -1,6 +1,6 @@
 import {EventDispatcher, HemisphereLight, DirectionalLight, Vector3} from 'three';
 import {EVENT_UPDATED} from '../core/events.js';
-import {renderProfile, isStudio} from './render_profile.js';
+import {renderProfile, isStudio} from '../core/render_profile.js';
 
 export class Lights extends EventDispatcher
 {